Backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pandas>=2.2.1
numpy>=1.26.4
scikit-learn>=1.4.0
scipy>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.1